ChatJLT/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-chroma==0.1.0
chromadb==0.4.22
openai==1.6.1
tiktoken==0.7.0
beautifulsoup4==4.12.3
aiohttp==3.9.1
pypdf==3.17.4
requests==2.31.0
pydantic==2.5.3
